all tests pass, except syntax, TBD

+++ test/test_syntax.jl
@@ -11,7 +11,7 @@ n = 1000
 nbeta = 10 
 X = [ones(n) randn((n, nbeta-1))] 
 beta0 = randn((nbeta,))
-Y = rand(n) .< ( 1 ./ (1. + exp(X * beta0)))
+Y = float64(rand(n) .< ( 1 ./ (1. + exp(X * beta0))))
 
 # define model
 ex = quote
